Fix 64-bit save/load: take intptr_t handle, not truncated int
SaveTheGame/LoadTheGame took 'int file' and the callers passed
'(int)(intptr_t)fp', truncating 64-bit FILE* pointers to 32 bits. The
function then cast the int back to FILE*, producing a sign-extended
garbage pointer on Win64 (observed as 0xFFFFFFFFB3E92190). Any save
or load whose fp happened to land outside the lower 4 GiB crashed.

Change the public signature to intptr_t and update all callers. Also
adds a --test-saveload self-test that does a save/load/save round-trip
and byte-compares the two saves. After the cast fix the test produces
identical 52361-byte outputs except for 3 bytes of uninitialized
struct padding inside objtype entries (heap-dependent and behaviorally
inert).

+++ src/wl_def.h
@@ -727,8 +727,8 @@ void 		CalcProjection (long focal);
 boolean		SetViewSize (unsigned width, unsigned height);
 void		NewGame (int difficulty,int episode);
 void 		NewViewSize (int width);
-boolean 	LoadTheGame(int file,int x,int y);
-boolean		SaveTheGame(int file,int x,int y);
+boolean 	LoadTheGame(intptr_t file,int x,int y);
+boolean		SaveTheGame(intptr_t file,int x,int y);
 void 		ShowViewSize (int width);
 void		ShutdownId (void);
